easter-egg: move to modules/wapuu-world, rename files to kebab-case
- Move src/components/easter-egg/ → src/modules/wapuu-world/
- Move src/assets/easter-egg/ → src/modules/wapuu-world/assets/
- Move src/hooks/use-konami-code.ts → src/modules/wapuu-world/use-konami-code.ts
- Rename EasterEggGame.tsx → wapuu-world-game.tsx, export WapuuWorldGame
- Rename Redux state/actions: isEasterEggOpen → isWapuuWorldOpen, openEasterEgg → openWapuuWorld, closeEasterEgg → closeWapuuWorld
- Update all imports in app.tsx and top-bar.tsx
