+++ apps/studio/src/modules/wapuu-world/assets/SPRITES.md
