+++ apps/studio/src/modules/wapuu-world/assets/wapuu-bg-far.png
