+++ apps/studio/src/modules/wapuu-world/assets/wapuu-bg-near.png
