+++ apps/studio/src/modules/wapuu-world/assets/wapuu-gutenberg-sprite.png
