+++ apps/studio/src/modules/wapuu-world/assets/wapuu-player-idle-sprite.png
