+++ apps/studio/src/modules/wapuu-world/assets/wapuu-player-sprite.png
