+++ apps/studio/src/modules/wapuu-world/assets/wapuu-sprites-coin.png
