+++ apps/studio/src/modules/wapuu-world/assets/wapuu-sprites-tiles.png
