+++ apps/studio/src/modules/wapuu-world/assets/wapuu-sprites.png
