+++ apps/studio/src/modules/wapuu-world/engine/game-loop.ts
